Polish XDG Base Directory implementation.
Prior to this commit, there was still some inconsistency in how XDG Base
Directory was handled. For instance, hooks were still being loaded from
the old-style config directory.

This commit unifies the handling of directory locations in the method
`init_dirs`. Directories are not hardcoded, respecting the XDG_*
environment variables, only falling back to the defaults if they are not
set. Any missing directories are created. Any tildes in the paths are
immediately expanded using `path.expanduser`.

For backward compatibility, the implementation prefers the old-style
directory if it exists, but emits a deprecation warning to the user on
stderr.

All methods should now use the `Config` instance variables `_config_dir`
and `_cache_dir` for the config and cache directories (respectively),
instead of calculating those paths themselves.