Restructure config system.
Instead of specifying defaults by directly assigning them to Pymodoro's
internal state, a PymodoroConfigParser (a derived class of
RawConfigParser) instance is constructed containing the defaults.
A separate function then takes such an instance and assigns its settings
to the internal state.

This allows us to easily merge user settings from the config file over
the defaults. It also obviates the need for a separate function for
creating the config file, because we can simply use
`PymodoroConfigParser.write`.

In addition, it provides the following benefits:

- Lays groundwork for supporting multiple configuration files of
  increasing priority (e.g. system-wide, followed by user
  configuration), if wanted.
- The config parser now tolerates missing options in the config file.
- Self-documenting: when the config file is missing, it will get created
  with all available settings populated with their default values.
- Hooks were made configurable.
- Allows all sounds used by Pymodoro to be customized via the config
  file.