Restructure state calculation algorithm to fix hooks.
Previously the state calculation algorithm worked by separately determining the
current state and then predicting (based on the current state and parameters
such as pomodoro and break lengths) what the next state will be, after the next
tick elapses. Unfortunately this method is unable to predict user actions and
will therefore never predict that the next state would be the active state
(because it can never predict a user clicking to start a pomodoro). This left
start hooks completely broken.

Fortunately, this kind of prediction is entirely unnecessary. The new algorithm
works as follows:

1. Initialize current state to IDLE_STATE.
2. Loop:

    a. Calculate the next state by comparing the mtime of the cache file with
       the current time. *Do not* assign this state as the current state yet.
    b. Run events (notifications, hooks) based on the difference between the current and next states.
    c. Assume the next state as the current state.
    d. Print output and play tick sounds.
    e. Wait for the next tick.
    f. Go back to a.