Add pomodoro end hook.
This is a hook which fires whenever a pomodoro is ended, regardless
of whether it because the pomodoro completed naturally or it was
terminated by the user.

A possible use case is to be able to pause system notifications when the
pomodoro starts and turn them back on when it is no longer active.