Actually source hooks from settings, but default them to empty strings.
There's no reason to specify default paths for the hooks in the config
file since the hooks are not created automatically.

Furthermore, since the only constraint on hooks is that they should be
executable files, there's no reason to suggest that they should be
Python modules (as was previously the case due to the default filenames
having the .py extension).